Moved the IGCSE and IAL subject list out of template to a different route to facilitate ease for future addition of different boards
