+++ src/app/IAL/layout.jsx
@@ -2,6 +2,7 @@ export const metadata = {
   title: "Subjects - IAL"
 };
 
+import {Home} from '@/components/homenav'
 export default function RootLayout({ children }) {
-  return (<>{children}</>)
+  return (<><Home showExtra/> {children}</>)
 }
