+++ src/app/IGCSE/layout.jsx
@@ -2,6 +2,7 @@ export const metadata = {
   title: "Subjects - IGCSE"
 };
 
+import {Home} from '@/components/homenav'
 export default function RootLayout({ children }) {
-  return (<>{children}</>)
+  return (<><Home showExtra/> {children}</>)
 }
