+++ src/app/IGCSE/page.jsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { supabase } from './lib/supabaseClient';
+import { supabase } from '../subjects/template/IGCSE/lib/supabaseClient';
 import Link from 'next/link';
 import { useEffect, useState } from 'react';
