--- src/app/subjects/psychology/IAL/layout.jsx
@@ -1,7 +1,0 @@
-export const metadata = {
-  title: "Subjects - IAL"
-};
-
-export default function RootLayout({ children }) {
-  return (<>{children}</>)
-}
