--- src/app/subjects/psychology/IGCSE/layout.jsx
@@ -1,7 +1,0 @@
-export const metadata = {
-  title: "Subjects - IGCSE"
-};
-
-export default function RootLayout({ children }) {
-  return (<>{children}</>)
-}
